Refactor quality control
