+++ Snakefile
@@ -5,150 +5,91 @@ scenarios = glob_wildcards("inputs/conf/{scenario}.json").scenario
 
 rule all:
     input:
-        expand(f'outputs/{{scenario}}/map_target2_drop_iqr_{config["iqr_scale"]}.parquet', scenario=scenarios)
+        expand('outputs/{scenario}/map_{plate_type}_drop_iqr.parquet', scenario=scenarios, plate_type=['prod', 'target2'])
 
-rule to_parquet:
+rule write_parquet:
     input: 'inputs/conf/{scenario}.json'
     output: 'outputs/{scenario}/raw.parquet'
-    run: qc.write_parquet(*input, *output)
+    run: qc.io.write_parquet(*input, *output)
 
-rule get_negcon_stats:
+rule compute_negcon_stats:
     input: 'outputs/{scenario}/raw.parquet'
     output: 'outputs/{scenario}/neg_stats.parquet'
-    run: qc.write_negcon_stats(*input, *output)
+    run: qc.stats.compute_negcon_stats(*input, *output)
 
-rule get_variant_feats:
-    input: 'outputs/{scenario}/neg_stats.parquet'
+rule select_variant_feats:
+    input: 'outputs/{scenario}/raw.parquet', 'outputs/{scenario}/neg_stats.parquet'
     output: 'outputs/{scenario}/variant_feats.parquet'
-    run: qc.write_variant_features(*input, *output)
+    run: qc.stats.select_variant_features(*input, *output)
 
-rule normalize_features:
+rule mad_normalize:
     input:
-        'outputs/{scenario}/raw.parquet',
+        'outputs/{scenario}/variant_feats.parquet',
         'outputs/{scenario}/neg_stats.parquet',
-        'outputs/{scenario}/variant_feats.parquet'
-    output: 'outputs/{scenario}/normalized.parquet'
+    output: 'outputs/{scenario}/mad.parquet'
     run:
-        qc.write_normalize_features(*input, *output)
+        qc.normalize.mad(*input, *output)
 
-rule norm_feat_stats:
+rule compute_norm_stats:
     input:
-        'outputs/{scenario}/normalized.parquet',
+        'outputs/{scenario}/mad.parquet',
     output:
-        'outputs/{scenario}/norm_feat_stats.parquet'
+        'outputs/{scenario}/norm_stats.parquet'
     run:
-        qc.write_normalize_feat_stats(*input, *output)
+        qc.stats.compute_stats(*input, *output)
 
 rule iqr_outliers:
     input:
-        'outputs/{scenario}/normalized.parquet',
-        'outputs/{scenario}/norm_feat_stats.parquet'
+        'outputs/{scenario}/mad.parquet',
+        'outputs/{scenario}/norm_stats.parquet'
     output:
-        'outputs/{scenario}/iqr_outliers_' f'{config["iqr_scale"]}.parquet'
+        'outputs/{scenario}/iqr_outliers.parquet'
     run:
-        qc.write_iqr_outliers(config['iqr_scale'], *input, *output)
+        qc.outliers.iqr(config['iqr_scale'], *input, *output)
 
-rule map_production_drop_iqr_outliers:
+rule drop_outlier_cols:
     input:
-        'outputs/{scenario}/normalized.parquet',
-        'outputs/{scenario}/variant_feats.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
+        'outputs/{scenario}/mad.parquet',
+        'outputs/{scenario}/iqr_outliers.parquet'
     output:
-        f'outputs/{{scenario}}/ap_prod_drop_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_prod_drop_iqr_{config["iqr_scale"]}.parquet'
+        'outputs/{scenario}/mad_drop_outlier_cols.parquet'
+    run:
+        qc.outliers.drop_cols(*input, *output)
+
+rule ap_production_drop_iqr_outliers:
+    input:
+        'outputs/{scenario}/mad_drop_outlier_cols.parquet'
+    output:
+        'outputs/{scenario}/ap_prod_drop_iqr.parquet',
     params:
         plate_types = ['COMPOUND'],
         min_replicates = 2,
         max_replicates = 100 # POSCONs and DMSO have a lot more
     run:
-        qc.write_map_drop_outlier_cols(*input, *output, **params)
+        qc.evaluation.average_precision(*input, *output, **params)
 
-rule map_target2_drop_iqr_outliers:
+rule map_production_drop_iqr_outliers:
     input:
-        'outputs/{scenario}/normalized.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
+        'outputs/{scenario}/ap_prod_drop_iqr.parquet'
     output:
-        f'outputs/{{scenario}}/ap_target2_drop_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_target2_drop_iqr_{config["iqr_scale"]}.parquet'
+        'outputs/{scenario}/map_prod_drop_iqr.parquet'
+
+rule ap_target2_drop_iqr_outliers:
+    input:
+        'outputs/{scenario}/mad_drop_outlier_cols.parquet'
+    output:
+        'outputs/{scenario}/ap_target2_drop_iqr.parquet'
     params:
         plate_types = ['TARGET2'],
         min_replicates = 2,
         max_replicates = float('inf')
     run:
-        qc.write_map_drop_outlier_cols(*input, *output, **params)
+        qc.evaluation.average_precision(*input, *output, **params)
 
-rule map_production_clip_iqr_outliers:
+rule map_target2_drop_iqr_outliers:
     input:
-        'outputs/{scenario}/normalized.parquet',
-        'outputs/{scenario}/variant_feats.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
+        'outputs/{scenario}/ap_target2_drop_iqr.parquet'
     output:
-        f'outputs/{{scenario}}/ap_prod_clip_{config["clip_value"]}_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_prod_clip_{config["clip_value"]}_iqr_{config["iqr_scale"]}.parquet'
-    params:
-        plate_types = ['COMPOUND'],
-        min_replicates = 2,
-        max_replicates = 100, # POSCONs and DMSO have a lot more
-        clip_value = config['clip_value']
+        'outputs/{scenario}/map_target2_drop_iqr.parquet'
     run:
-        qc.write_map_clip_outlier_cols(*input, *output, **params)
-
-rule map_target2_clip_iqr_outliers:
-    input:
-        'outputs/{scenario}/normalized.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
-    output:
-        f'outputs/{{scenario}}/ap_target2_clip_{config["clip_value"]}_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_target2_clip_{config["clip_value"]}_iqr_{config["iqr_scale"]}.parquet'
-    params:
-        plate_types = ['TARGET2'],
-        min_replicates = 2,
-        max_replicates = float('inf'),
-        clip_value = config['clip_value']
-    run:
-        qc.write_map_clip_outlier_cols(*input, *output, **params)
-
-
-rule map_target2_impute_median_iqr_outliers:
-    input:
-        'outputs/{scenario}/normalized.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
-    output:
-        f'outputs/{{scenario}}/ap_target2_impute_median_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_target2_impute_median_iqr_{config["iqr_scale"]}.parquet'
-    params:
-        plate_types = ['TARGET2'],
-        min_replicates = 2,
-        max_replicates = float('inf'),
-    run:
-        qc.write_map_impute_median_outlier_cols(*input, *output, **params)
-
-
-rule map_target2_impute_knn_iqr_outliers:
-    input:
-        'outputs/{scenario}/normalized.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
-    output:
-        f'outputs/{{scenario}}/ap_target2_impute_knn_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_target2_impute_knn_iqr_{config["iqr_scale"]}.parquet'
-    params:
-        plate_types = ['TARGET2'],
-        min_replicates = 2,
-        max_replicates = float('inf'),
-    run:
-        qc.write_map_impute_knn_outlier_cols(*input, *output, **params)
-
-
-rule map_target2_rank_int_drop_iqr_outliers:
-    input:
-        'outputs/{scenario}/normalized.parquet',
-        f'outputs/{{scenario}}/iqr_outliers_{config["iqr_scale"]}.parquet'
-    output:
-        f'outputs/{{scenario}}/ap_target2_rank_int_drop_iqr_{config["iqr_scale"]}.parquet',
-        f'outputs/{{scenario}}/map_target2_rank_int_drop_iqr_{config["iqr_scale"]}.parquet'
-    params:
-        plate_types = ['TARGET2'],
-        min_replicates = 2,
-        max_replicates = float('inf'),
-    run:
-        qc.write_map_rank_int_drop_outlier_cols(*input, *output, **params)
+        qc.evaluation.mean_average_precision(*input, *output)
